fixed bug(data type)

+++ src/main/java/telran/probes/dto/DeviationData.java
@@ -4,6 +4,6 @@ import jakarta.validation.constraints.Min;
 import telran.probes.messages.ErrorMessages;
 
 
-public record DeviationData(@Min(value= 1, message= ErrorMessages.WRONG_SENSOR_ID) int id, double deviation, double value, long timestamp) {
+public record DeviationData(@Min(value= 1, message= ErrorMessages.WRONG_SENSOR_ID) long id, double deviation, double value, long timestamp) {
 
 }
